+++ src/helpers/nullAsUndefined.internal.ts
@@ -1,0 +1,2 @@
+export const nullAsUndefined = <T>(prop: T): NonNullable<T> | undefined =>
+  prop === null ? undefined : (prop as NonNullable<T>);